phy/lpddr4: move common utils up a directory
